+++ encryption_machine/encryption/utils/qr_code.py
